pyproject.toml
[tool.poetry]
name = "group_to_channel"
version = "0.1.0"
description = ""
authors = ["Zhao Liang"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.0"
qrcode = "^6.1"
pillow = "^7.2.0"
requests = "^2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
